requirements.txt
altgraph==0.17
aniso8601==8.0.0
cefpython3==66.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
future==0.18.2
gevent==20.6.1
greenlet==0.4.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.4
pefile==2019.4.18
Pillow==7.1.2
pycparser==2.20
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
wxPython==4.1.0
zope.event==4.4
zope.interface==5.1.0
